x/coin: move things out of the way so it builds
