+++ x/auth/decorator.go_
@@ -1,0 +1,37 @@
+package auth
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk"
+	"github.com/cosmos/cosmos-sdk/store"
+	"github.com/cosmos/cosmos-sdk/types"
+)
+
+type CheckSignatures struct{}
+
+var _ types.Decorator = CheckSignatures{}
+
+// CheckTx verifies the signatures are correct - fulfills Middlware interface
+func (CheckSignatures) CheckTx(ctx sdk.Context, store store.MultiStore,
+	tx interface{}, next sdk.Checker) (res sdk.CheckResult, err error) {
+
+	// Check that signatures match
+	// TODO
+
+	// Add info to context
+	// TODO
+
+	return next.CheckTx(ctx2, store, tx)
+}
+
+// DeliverTx verifies the signatures are correct - fulfills Middlware interface
+func (CheckSignatures) DeliverTx(ctx sdk.Context, store store.MultiStore,
+	tx interface{}, next sdk.Deliverer) (res sdk.DeliverResult, err error) {
+
+	// Check that signatures match
+	// TODO
+
+	// Add info to context
+	// TODO
+
+	return next.DeliverTx(ctx2, store, tx)
+}
